requirements.txt
appname==100
asgiref==3.4.1
colorama==0.4.4
Django==3.2.6
django-configurations==2.2
django-crispy-forms==1.12.0
django-filter==2.4.0
djangorestframework==3.12.4
loguru==0.5.3
Pillow==8.3.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
win32-setctime==1.0.3